adding client for the symfony built-in reverse proxy HttpCache
